requirements.txt
# FastAPI 및 웹 서버
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 데이터베이스
psycopg2-binary>=2.9.9

# 이미지 처리
Pillow>=10.0.0

# PDF 처리
PyMuPDF>=1.23.0

# OpenAI API
openai>=1.0.0

# 벡터 검색 (RAG)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# 환경 변수 관리
python-dotenv>=1.0.0

# 유틸리티
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Google Gemini (선택사항)
google-generativeai>=0.3.0

# Upstage API (선택사항)
# upstage>=0.1.0  # 필요시 주석 해제

# 스케줄러
APScheduler>=3.10.0
